Create pipeline to auto bump packages
Introduce automatic bumping of ruby, rubygems and libyaml using the dynamic
metalink concourse resource. The concept is copied from the golang-release.

The pipeline can be found here:
https://bosh-ci.cpi.sapcloud.io/teams/main/pipelines/update-ruby-release

It includes the following tasks:
Test - Smoke test which deploys a dummy deployment with bumped packages
and runs an errand currently verifying the rubygems version.
Bump - Bumps blobs of BOSH release and uploads them to s3 blobstore.
Finalize - Creates final BOSH release, uploads it to s3 blobstore and
creates a new version tag for the ruby-release.

[#161676998](https://www.pivotaltracker.com/story/show/161676998)

Co-authored-by: Kai Hofstetter <kai.hofstetter@sap.com>
Co-authored-by: Florian Nachtigall <florian.nachtigall@sap.com>