Remove patch versions from package names
This will make it easier for releases to auto-bump the minor version of ruby
they want as the paths to the packages and compile/runtime scripts will be stable

Also removed some interesting "cpi" code that seems like it was treating
the ruby-release as a cpi for the purposes of tricking the bosh cli
into compiling it on darwin. Doesn't seem like this was used anymore.

[#182075817]