Ruby package is compatible with ARM-based Macs
There's one case where the Ruby package is compiled on a Mac, and that's
when the `bosh create-env` command is run on a Mac that's targeting a
vCenter. The BOSH vSphere CPI depends on this Ruby package, attempts to
compile it, and fails (on ARM-based Macs).

This commit fixes that by detecting homebrew-installed OpenSSL & setting
the appropriate compilation options.

We dispensed with the `--with-opt-dir` flag. It didn't seem necessary,
and couldn't determine why it was initially included.

Fixes, during `bosh create-env`:
```
error: use of undeclared identifier 'RUBY_FUNCTION_NAME_STRING'
        rp(ci);
```

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>
Signed-off-by: Joseph Palermo <jpalermo@vmware.com>