Ensure git-lfs is installed for bumping gems
Some projects vendor their gems into the repo, but do so under a
directory controlled by `git-lfs`. If git-lfs is not installed when
bumping such projects, you can get hard-to-decipher errors like:
```
ArgumentError: "701b4d89" is not an octal string
```
because the existing gems were not translated from their git lfs
reference to the actual file.