Remove CFLAGS when compiling ruby
Bosh director template rendering runs about 2x as slow when the CFLAGS are specified and GCC is used.

This is likely due to the fact that when you specify CFLAGS, it removes the default "-O3"
flag: https://github.com/cloudfoundry/bosh-linux-stemcell-builder/issues/408#issuecomment-2625120616

Since there isn't a clear reason why we are configuring "-fPIC" anyway, lets just remove them
as we are planning on removing CLang from Noble stemcells and we'll be back to using GCC again.